Refactored QC tests, split tests & fixtures into seperate files.
